Add support for mf and mfuuid static libraries
Added static libraries 'mf' and 'mfuuid' to the linked Rust libraries in build.rs. This change is necessary to provide the required multimedia capabilities in the project. Additionally, created a new file initguid.cpp to allow for the direct inclusion of the Enhanced Video Renderer (EVR) header and modified the CMakeLists.txt file to include the newly created file during the build process.